+++ COLLADA2GLTF/extensions/o3dgc-compression/GLTF-Open3DGC.cpp
@@ -265,7 +265,7 @@ namespace GLTF
         O3DGCSC3DMCPredictionMode jointPrediction = _predictionModeForString(config->stringForKeyPath("extensions.Open3DGC.quantization.JOINT"));
         
         GLTFOutputStream *outputStream = asset->createOutputStreamIfNeeded(kCompressionOutputStream).get();
-        size_t bufferOffset = outputStream->length();
+        int bufferOffset = (int)outputStream->length();
         
         O3DGCSC3DMCPredictionMode floatAttributePrediction = O3DGC_SC3DMC_PARALLELOGRAM_PREDICTION;
         
@@ -316,7 +316,7 @@ namespace GLTF
         ifs.SetCoordIndex((unsigned short * const ) allConcatenatedIndices);
         ifs.SetIndexBufferID(primitiveIDs);
         
-        size_t vertexCount = 0;
+        long vertexCount = 0;
         
         std::vector <GLTF::Semantic> semantics = mesh->allSemantics();
         for (unsigned int i = 0 ; i < semantics.size() ; i ++) {
@@ -326,8 +326,8 @@ namespace GLTF
             
             for (size_t j = 0 ; j < attributesCount ; j++) {
                 shared_ptr <GLTFAccessor> meshAttribute = mesh->getMeshAttribute(semantic, j);
-                vertexCount = meshAttribute->getCount();
-                size_t componentsPerElement = meshAttribute->componentsPerElement();
+                vertexCount = (long)meshAttribute->getCount();
+                long componentsPerElement = (long)meshAttribute->componentsPerElement();
                 char *buffer = (char*)meshAttribute->getBufferView()->getBufferDataByApplyingOffset();
                 switch (semantic) {
                     case POSITION:
@@ -405,7 +405,7 @@ namespace GLTF
         BinaryStream bstream(vertexCount * 8);
         SC3DMCEncoder <unsigned short> encoder;
         shared_ptr<JSONObject> compressedData(new JSONObject());
-        compressedData->setInt32("verticesCount", vertexCount);
+        compressedData->setInt32("verticesCount", (int)vertexCount);
         compressedData->setInt32("indicesCount", allIndicesCount);
         //Open3DGC binary is disabled
         params.SetStreamType(CONFIG_STRING(asset, "compressionMode") == "binary" ? O3DGC_STREAM_TYPE_BINARY : O3DGC_STREAM_TYPE_ASCII);
@@ -440,7 +440,7 @@ namespace GLTF
         }
     }
     
-    void encodeDynamicVector(float *buffer, const std::string &path, size_t componentsCount, size_t count, GLTFAsset* asset) {
+    void encodeDynamicVector(float *buffer, const std::string &path, long componentsCount, long count, GLTFAsset* asset) {
         GLTFOutputStream *outputStream = asset->createOutputStreamIfNeeded(kCompressionOutputStream).get();
         Real max[32];
         Real min[32];
@@ -532,7 +532,7 @@ namespace GLTF
                                                              const std::string& parameterSID,
                                                              const std::string& accessorUID,
                                                              const std::string& parameterType,
-                                                             unsigned char* buffer, size_t byteLength,
+                                                             unsigned char* buffer, int byteLength,
                                                              bool isInputParameter,
                                                              GLTFAsset *asset) {
         //setup
@@ -540,7 +540,7 @@ namespace GLTF
         shared_ptr <GLTF::JSONObject> accessors = asset->root()->createObjectIfNeeded(kAccessors);
         shared_ptr<JSONObject> parameter(new JSONObject());
         unsigned int glType = profile->getGLenumForString(parameterType);
-        parameter->setUnsignedInt32("count", cvtAnimation->getCount());
+        parameter->setUnsignedInt32("count", (int)cvtAnimation->getCount());
         parameter->setString(kType, profile->getTypeForGLType(glType));
         parameter->setUnsignedInt32(kComponentType, profile->getGLComponentTypeForGLType(glType));
 
@@ -548,18 +548,18 @@ namespace GLTF
         cvtAnimation->parameters()->setString(parameterSID, accessorUID);
         
         //write
-        size_t byteOffset = 0;
+        int byteOffset = 0;
         bool shouldEncodeOpen3DGC = CONFIG_STRING(asset, "compressionType")  == "Open3DGC";
         GLTFOutputStream *outputStream = shouldEncodeOpen3DGC ? asset->createOutputStreamIfNeeded(kCompressionOutputStream).get() : asset->createOutputStreamIfNeeded(asset->getSharedBufferId()).get();;
-        byteOffset = outputStream->length();
+        byteOffset = (int)outputStream->length();
         parameter->setUnsignedInt32("byteOffset", byteOffset);
         
         if (shouldEncodeOpen3DGC) {
-            size_t componentsCount = profile->getComponentsCountForType(parameter->getString(kType));
+            long componentsCount = (long)profile->getComponentsCountForType(parameter->getString(kType));
             if (componentsCount) {
-                encodeDynamicVector((float*)buffer, parameterSID, componentsCount, cvtAnimation->getCount(), asset);
+                encodeDynamicVector((float*)buffer, parameterSID, componentsCount, (long)cvtAnimation->getCount(), asset);
                 
-                byteLength = outputStream->length() - byteOffset;
+                byteLength = (int)outputStream->length() - byteOffset;
                 
                 shared_ptr<JSONObject> extensionsObject = parameter->createObjectIfNeeded(kExtensions);
                 shared_ptr<JSONObject> compressionObject = extensionsObject->createObjectIfNeeded("Open3DGC-compression");
@@ -589,7 +589,7 @@ namespace GLTF
     void setupAndWriteAnimationParameter(GLTFAnimation *cvtAnimation,
                                          const std::string& parameterSID,
                                          const std::string& parameterType,
-                                         unsigned char* buffer, size_t byteLength,
+                                         unsigned char* buffer, int byteLength,
                                          bool isInputParameter,
                                          GLTFAsset *asset) {
